Small memory improvement: BObjectImp.type_ is now u8 instead of int
It is very unlikely to grow to more than 256 elements any time soon.
This should save 3 bytes for every single object that derived from it.
Also added explicit int conversions and a not, for better robustness.

+++ pol-core/bscript/berror.h
@@ -40,7 +40,7 @@ namespace Pol {
 	  virtual char packtype() const POL_OVERRIDE;
 	  virtual const char* typetag() const POL_OVERRIDE;
 	  virtual const char* typeOf() const POL_OVERRIDE;
-	  virtual int typeOfInt() const POL_OVERRIDE;
+	  virtual u8 typeOfInt() const POL_OVERRIDE;
 
 	  virtual bool isEqual( const BObjectImp& objimp ) const POL_OVERRIDE;
 	  virtual bool isTrue() const POL_OVERRIDE;
