+++ pol-core/pol/module/npcmod.cpp
@@ -154,7 +154,7 @@ namespace Pol {
       BoundingBoxObjImp( ) : BApplicObj<Mobile::BoundingBox>( &bounding_box_type ) {}
       explicit BoundingBoxObjImp( const Mobile::BoundingBox& b ) : BApplicObj<Mobile::BoundingBox>( &bounding_box_type, b ) {}
 	  virtual const char* typeOf() const POL_OVERRIDE { return "BoundingBox"; }
-	  virtual int typeOfInt() const POL_OVERRIDE { return OTBoundingBox; }
+	  virtual u8 typeOfInt() const POL_OVERRIDE { return OTBoundingBox; }
 	  virtual BObjectImp* copy() const POL_OVERRIDE { return new BoundingBoxObjImp( value() ); }
 
 	};
